BUG: Fix initialization of DataFrame from dict with NaN as key (#18600)
closes #18455
closes #19646

* TST: removed workaround for #18455

* TST: asv for DataFrame from dict with index/columns

+++ pandas/tests/io/test_excel.py
@@ -461,7 +461,7 @@ class ReadingTestsBase(SharedItems):
             )
         expected_header_none = DataFrame(pd.Series([0], dtype='int64'))
         tm.assert_frame_equal(actual_header_none, expected_header_none)
-        expected_header_zero = DataFrame(columns=[0], dtype='int64')
+        expected_header_zero = DataFrame(columns=[0])
         tm.assert_frame_equal(actual_header_zero, expected_header_zero)
 
     @td.skip_if_no('openpyxl')
